==> data_preparation/01_create_csv.ipynb <==
### Erstellung der CSV Datei / Befüllen der Spalten

* die hier aufgerufenen CSV Dateien "img.csv", "img2.csv", "bilder_neu.csv", "sharepics.csv" wurden im Verlauf der Arbeit umbenannt und im Projektverzeichnis verschoben. DSie sind nicht mehr Teil des Projekts, aber waren die Grundlage für die finale Datei "images_data_with_groundtruth.csv

#### Erstellen der Spalten und befüllen mit Bildname, Bildordnername (Kategorie)

# dieser Codeblock wurde von ChatGPT generiert und nur leicht modifiziert
import os
import csv


images = "./bilder"  

# Ziel-CSV-Datei
csv_datei = "img.csv"

# Bildzähler
zaehler = 1

# CSV vorbereiten
with open(csv_datei, mode='w', newline='', encoding='utf-8') as csv_file:
    csv_writer = csv.writer(csv_file)
    csv_writer.writerow(["orignal name", "new name", "catgeory"])

    # Durchlaufe alle Unterordner
    for root, dirs, files in os.walk(images):
        for datei in files:
            if datei.lower().endswith((".jpg", ".jpeg", ".png")):
                alter_pfad = os.path.join(root, datei)
                relative_pfad = os.path.relpath(alter_pfad, images)

                # Kategorie = Name des Unterordners (direkt unter "bilder")
                kategorie = os.path.relpath(root, images).split(os.sep)[0]

                # Neuer Dateiname: nur Nummer mit gleicher Endung
                dateiendung = os.path.splitext(datei)[1].lower()
                neuer_dateiname = f"{zaehler:04d}{dateiendung}"
                neuer_pfad = os.path.join(root, neuer_dateiname)
                relative_neu = os.path.relpath(neuer_pfad, images)

                # Umbenennen
                os.rename(alter_pfad, neuer_pfad)


                # In CSV speichern
                csv_writer.writerow([relative_pfad, os.path.relpath(neuer_pfad, images), kategorie])
                zaehler += 1





import pandas as pd
import os

file_path = os.path.join( "img.csv")


df = pd.read_csv(file_path)

df.head(10)

#### Befüllen der Spalte "new name" 

import pandas as pd
import os

df = pd.read_csv("img2.csv", sep=",") 

# Nur den Dateinamen extrahieren (z. B. bild_0001.jpg -> 0001.jpg)
df["new name"] = df["new name"].apply(lambda x: os.path.basename(x).replace("bild_", ""))

df.to_csv("img_2.csv", index=False, sep=",")


file_path = os.path.join( "img_2.csv")


df = pd.read_csv(file_path)

df.tail(3)

#### Erstellen weiterer Spalten

df['URL'] = ''
df['Kanal'] = ''

df['channel name'] = ''
df.rename(columns={'Kanal': 'Plattform'}, inplace=True)

df.head(3)


#### manuelles Eintragen der Werte im Dataframe

neue_zeile = pd.DataFrame([{
    'original name': 'Die Berliner Allee wude schick gemacht!Am 6. Oktober 2024auf die Straße!Stoppt den III. Weg!Für ein Ende der Nazi-Kampftrainings im Sportkplex in der Weißenseer Rennbahnstraße!Für die befreite Geschafft!Wir sehen un.jpg',
    'new name' : '0242.jpg',
    'category' : 'abfotografie',
    'URL': 'https://www.instagram.com/p/DAjFfRYMfBE/?img_index=1',
    'Plattform': 'Instagram',
    'channel name' : 'antifanordost'
}])

df = pd.concat([df, neue_zeile], ignore_index=True)


#print(df.columns.tolist())


#df.rename(columns={
#    'orignal name': 'original name',
#    'catgeory': 'category'
#}, inplace=True)


#df = df.drop(index=193)


df.tail(2)

# Entfernt doppelte Spaltennamen – behält nur die erste Version
#df = df.loc[:, ~df.columns.duplicated()]


#print(df.columns)
#df.tail()


df[df['new name'] == '0121.jpg'].index


df.loc[71]

df.loc[df['original name'] == 'fancy_typo/bild_0077.jpg', ['original name', 'URL', 'Plattform', 'channel name']] = [
    'Bitte kommt!.jpg',
    'https://www.instagram.com/p/C3JUclUNKyq/',
    'Instagram',
    'omasgegenrechts'
]




df.to_csv("bilder_neu.csv")


import pandas as pd

df_neu = pd.read_csv("bilder_neu.csv")  
df_neu.head()


df[df_neu["Plattform"].isna()]


df_neu = df_neu[df_neu["Plattform"].notna()]



df_neu["Plattform"].isna().sum()


df_neu[df_neu["Plattform"].isna()]

df_neu = df_neu.reset_index(drop=True)
df_neu.to_csv("sharepics.csv")  


import pandas as pd

df_sharepics = pd.read_csv("sharepics.csv")  
df_sharepics.head()


#df = df_sharepics.loc[:, ~df_sharepics.columns.str.contains("^Unnamed")]


#df.head()

df.to_csv("sharepics.csv", index=False)


df_sharepics = pd.read_csv("sharepics.csv")  
df_sharepics.head()

import os

ordnerpfad = "bilder_neu"

for _, row in df.iterrows():
    original = row["original name"]
    neu = row["new name"]
    
    alter_pfad = os.path.join(ordnerpfad, original)
    neuer_pfad = os.path.join(ordnerpfad, neu)
    
    if os.path.exists(alter_pfad):
        os.rename(alter_pfad, neuer_pfad)
        print(f"{original} ➜ {neu}")
    else:
        print(f"Datei nicht gefunden: {original}")


df_sharepics[~df["new name"].str.match(r"^\d+\.(jpg|jpeg)$", na=False)]


df_sharepics.shape


df_sharepics[df["Plattform"] == "Telegram"]


df_sharepics[df["new name"] == "0200.jpg"]

df_sharepics.tail()

duplikate = df_sharepics[df.duplicated()]

print(duplikate)

duplikate_nach_name = df_sharepics[df.duplicated(subset=['new name'])]

print(duplikate_nach_name)

anzahl_pro_kategorie = df_sharepics['category'].value_counts()

print(anzahl_pro_kategorie)



#df_sharepics['category'] = df_sharepics['category'].replace('typo_aligment', 'typo_alignment')


#### Speicherung der der CSV Datei nach allen Eintragungen

df.to_csv("sharepics.csv", index=False)

df_sharepics = pd.read_csv("sharepics.csv")  
df_sharepics.head()
