==> data_preparation/02_enter_OCR_ChatGPT.ipynb <==
### Eintragung OCR-Text von ChatGPT in CSV

manuelle Eintragung pro Bild

* die hier benutzten CSV Dateien wurden im Verlauf der Arbeit umbenannt und im Projektverzeichnis verschoben. Sie sind nicht mehr Teil des Projekts, aber waren die Grundlage für die finale Datei "images_data_with_groundtruth.csv

import pandas as pd
import os
import csv
df = pd.read_csv("sharepics.csv")  
df.head()

print(df.loc[df['new name'] == '0025.jpg', 'OCR_ChatGPT4'].values)

df[~df['OCR_ChatGPT4'].isna()]


df.loc[df['new name'] == '0002.jpg', 'OCR_ChatGPT4'] = "SCHLIESST EUCH AN! KUNDGEBUNG #SOLIDARITÄT MIT ALLEN! BERICHTE: WIE WAR DIE BERLIN-DEMO WIRKLICH? DRESDEN, GROßER GARTEN, 8.8. UM 16 UHR"



print(df.loc[df['new name'] == '0002.jpg', 'OCR_ChatGPT4'].values)

df.to_csv("sharepics_withtext.csv", index=False)


df = pd.read_csv("sharepics_withtext.csv")  
df.head()

#df[~df['OCR_ChatGPT4'].isna()]

df.loc[df['new name'] == '0242.jpg', 'OCR_ChatGPT4'] = "SCHAUT NICHT WEG! Stoppt die Nazikampf-trainings in Weißensee 06.10.2024 DEMO 13 Uhr Antonplatz KEINEN RAUM DEM III. WEG! "



print(df.loc[df['new name'] == '0242.jpg', 'OCR_ChatGPT4'].values)

#df.loc[df['new name'] == '0216.jpg', 'text'] = "KRITISCHE PERSPEKTIVE AUF DEN OSTERMARSCH ’23 SAMSTAG 08.04. AB 13 UHR ▶ MÜLLERSTRASSE 146 Warum Linkssein.. ..nicht vor Konservatismus schützt.. HUFESENTHEORIE ?QUERFRONT? ..Die Hufeisentheorie aber trotzdem falsch ist. RIGHT TO RESIST GERADEDENKEN"



#print(df.loc[df['new name'] == '0216.jpg', 'text'].values)

#df['text'] = df['text'].astype(str)

#### "Auffüllen" der Spalte "text" mit dem OCR Output von ChatGPT

df["text"] = df["text"].fillna(df["OCR_ChatGPT4"])

df.to_csv("sharepics_withtext.csv", index=False)
