==> data_preparation/03_enter_ground_truth_csv.ipynb <==
### generierten Text von ChatGPT prüfen und manuell verbesserten Text in Spalte "text" als ground truth speichern.
* die hier aufgerufene CSV Datei "sharepics_withtext.csv" wurde im Verlauf der Arbeit umbenannt und im Projektverzeichnis verschoben. Die hier aufgerufenen CSV Datei ist nicht mehr Teil des Projekts, aber war die Grundlage für die finale Datei "images_data_with_groundtruth.csv
* die Ausführung des notebooks würde zu Fehlern führen - es dient nur zur Anschauung

import pandas as pd

df = pd.read_csv("sharepics_withtext.csv")
df.head()


df = df.rename(columns={'new name': 'name'})

df.to_csv('sharepics_withtext.csv', index=False)

df = pd.read_csv("sharepics_withtext.csv")
df.head()

print(df.loc[df['name'] == '0001.jpg', 'text'].values)

df.loc[df['name'] == '0001.jpg', 'text'] = 'SPRACHCAFE WEIHNACHTSFEIER DIENSTAG 21.12 OLOF-PALME ZENTRUM 19 UHR WIR FREUEN UNS AUF EUCH!'


df.to_csv('sharepics_withtext.csv', index=False)
