==> data_preparation/04_rename_columns.ipynb <==
### Erstellung der finalen CSV- Datei

import pandas as pd

df = pd.read_csv("../../sharepics_withtext.csv")  
df.head()

df.rename(columns={'name': 'file_name'}, inplace=True)

df.head(1)

df["category"] = df["category"].replace({
    "background": "noisy",
    "typo_alignment": "skewed",
    "fancy_typo": "fancy",
    "abfotografie": "photo"
})



df.head(1)

df.to_csv("data/image_data_with_groundtruth.csv", index=False)
